Оладка получения элемента по uuid.

+++ src/main/java/org/sber/service/testservice/repo/ClientProductRepo.java
@@ -3,5 +3,6 @@ package org.sber.service.testservice.repo;
 import org.sber.service.testservice.model.ClientProduct;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClientProductRepo extends JpaRepository<ClientProduct, Long> {
-}
+
+//public interface ClientProductRepo extends JpaRepository<ClientProduct, Long> {
+public interface ClientProductRepo extends JpaRepository<ClientProduct, Integer> {}
